+++ app/src/main/java/app/passwordstore/util/settings/PreferenceKeys.kt
@@ -17,7 +17,7 @@ object PreferenceKeys {
   )
   const val CLEAR_CLIPBOARD_20X = "clear_clipboard_20x"
   const val CLEAR_CLIPBOARD_HISTORY = "clear_clipboard_history"
-  @Deprecated(message = "We have retired persistent caching of passphrases")
+  @Deprecated(message = "We have retired persistent caching of passwords")
   const val CLEAR_SAVED_PASS = "clear_saved_pass"
   const val COPY_ON_DECRYPT = "copy_on_decrypt"
   const val ENABLE_DEBUG_LOGGING = "enable_debug_logging"
@@ -54,9 +54,7 @@ object PreferenceKeys {
 
   @Deprecated("Git branch is no longer stored in preferences")
   const val GIT_BRANCH_NAME = "git_branch"
-  @Deprecated(
-    message = "Now unused, as we have retired persistent caching of passphrases/passwords"
-  )
+  @Deprecated(message = "Now unused, as we have retired persistent caching of the https password")
   const val HTTPS_PASSWORD = "https_password"
   const val LENGTH = "length"
   const val OREO_AUTOFILL_CUSTOM_PUBLIC_SUFFIXES = "oreo_autofill_custom_public_suffixes"
@@ -75,12 +73,15 @@ object PreferenceKeys {
   const val SHOW_HIDDEN_CONTENTS = "show_hidden_contents"
   const val SORT_ORDER = "sort_order"
   const val SHOW_PASSWORD = "show_password"
+  const val UNLOCK_PASSWORDS_WITH_PIN = "unlock_passwords_with_pin"
   const val SSH_KEY = "ssh_key"
   const val SSH_KEYGEN = "ssh_keygen"
+
   @Deprecated(
-    message = "Now unused, as we have retired persistent caching of passphrases/passwords"
+    message = "Now unused, as we have retired persistent caching of the ssh key passphrase"
   )
   const val SSH_KEY_LOCAL_PASSPHRASE = "ssh_key_local_passphrase"
+
   const val SSH_OPENKEYSTORE_CLEAR_KEY_ID = "ssh_openkeystore_clear_keyid"
   const val SSH_OPENKEYSTORE_KEYID = "ssh_openkeystore_keyid"
   const val SSH_SEE_KEY = "ssh_see_key"
@@ -101,8 +102,9 @@ object PreferenceKeys {
   const val ASCII_ARMOR = "pgpainless_ascii_armor"
 
   @Deprecated(
-    message = "Now unused, as we have retired persistent caching of passphrases/passwords"
+    message = "We refactored persistent caching of the PGP passphrase and this is no longer used"
   )
   const val CLEAR_PASSPHRASE_CACHE = "pgpainless_auto_clear_passphrase_cache_screen_off"
+
   const val CACHE_PASSPHRASE = "cache_passphrase_until_screen_off"
 }